add delete subject api

+++ model/subject.go
@@ -58,6 +58,14 @@ func (model SubjectModel) DelSubject(subject *Subject) bool {
 	if subject.Id == 0 {
 		return false
 	}
+
+	// 科目下有题目的话不允许删除
+	qm := QuestionModel{DB: model.DB}
+	list, _ := qm.GetQuestionList(subject.Id, 1, 1)
+	if list != nil && len(list) > 0 {
+		return false
+	}
+
 	_, err := model.DB.Delete(subject)
 	return err == nil
 }
